Chore: Address different review items - Rename files from `optimize_1q_decomposition` to `optiize_1q_sequences`. - Fixed previous oversight that counted operations before the pass was ran.
