+++ crates/cext/src/transpiler/passes/optimize_1q_sequences.rs
